Scene, blinder and shadow drawn
Most of the preliminary setup of the scene is complete. Object is drawn that creates a blind spot while a shadow is drawn in its path that is only visible if the user changes the viewing angle. Shadow stops at road intersection and shows area where cyclists are invisible. Needs to be animated, most values are easy to change (Scene setup, angles ect)